# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/credits/validate": {
      "post": {
        "description": "Verifica se o usuário possui créditos suficientes para uma operação específica",
        "operationId": "CreditsController_validateCredits",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateCreditsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validação realizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateCreditsResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Validar créditos suficientes",
        "tags": [
          "🔵 Créditos"
        ]
      }
    },
    "/v1/credits/consume": {
      "post": {
        "description": "Consome uma quantidade específica de créditos do usuário para um serviço",
        "operationId": "CreditsController_consumeCredits",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumeCreditsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Créditos consumidos com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumeCreditsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Créditos insuficientes ou dados inválidos"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consumir créditos",
        "tags": [
          "🔵 Créditos"
        ]
      }
    },
    "/v1/credits/add": {
      "post": {
        "description": "Adiciona créditos a um usuário (apenas administradores)",
        "operationId": "CreditsController_addCredits",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCreditsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Créditos adicionados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddCreditsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "403": {
            "description": "Acesso negado - apenas administradores"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Adicionar créditos",
        "tags": [
          "🔵 Créditos"
        ]
      }
    },
    "/v1/credits/transfer": {
      "post": {
        "description": "Transfere créditos de um usuário para outro",
        "operationId": "CreditsController_transferCredits",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferCreditsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transferência realizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferCreditsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Créditos insuficientes ou dados inválidos"
          },
          "404": {
            "description": "Usuário origem ou destino não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Transferir créditos",
        "tags": [
          "🔵 Créditos"
        ]
      }
    },
    "/v1/credits/balance/{userId}": {
      "get": {
        "description": "Retorna o saldo atual de créditos de um usuário específico. Requer permissões de administrador.",
        "operationId": "CreditsController_getUserCreditBalance",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID único do usuário",
            "schema": {
              "example": "123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Saldo obtido com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "userId": {
                      "type": "string",
                      "example": "123"
                    },
                    "balance": {
                      "type": "number",
                      "example": 100
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter saldo de créditos de um usuário",
        "tags": [
          "🔵 Créditos"
        ]
      }
    },
    "/v1/credits/history/{userId}": {
      "get": {
        "description": "Obtém o histórico de transações de créditos de um usuário",
        "operationId": "CreditsController_getTransactionHistory",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID do usuário",
            "schema": {
              "example": "123",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Número máximo de transações a retornar",
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Número de transações a pular (paginação)",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico obtido com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreditTransaction"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Histórico de transações",
        "tags": [
          "🔵 Créditos"
        ]
      }
    },
    "/v1/credits/my-balance": {
      "get": {
        "description": "Retorna o saldo atual de créditos do usuário autenticado.",
        "operationId": "CreditsController_getMyCreditBalance",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Saldo obtido com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "userId": {
                      "type": "string",
                      "example": "123"
                    },
                    "balance": {
                      "type": "number",
                      "example": 100
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter meu saldo de créditos",
        "tags": [
          "🔵 Créditos"
        ]
      }
    },
    "/v1/credits/my-history": {
      "get": {
        "description": "Obtém o histórico de transações de créditos do usuário autenticado",
        "operationId": "CreditsController_getMyTransactionHistory",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Número máximo de transações a retornar",
            "schema": {
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Número de transações a pular (paginação)",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico obtido com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreditTransaction"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Meu histórico de transações",
        "tags": [
          "🔵 Créditos"
        ]
      }
    },
    "/v1/credits/my-consume": {
      "post": {
        "description": "Consome créditos do usuário autenticado",
        "operationId": "CreditsController_consumeMyCredits",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Créditos consumidos com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsumeCreditsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Créditos insuficientes ou dados inválidos"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consumir meus créditos",
        "tags": [
          "🔵 Créditos"
        ]
      }
    },
    "/v1/credits/my-validate": {
      "post": {
        "description": "Verifica se o usuário autenticado possui créditos suficientes",
        "operationId": "CreditsController_validateMyCredits",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Validação realizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateCreditsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Validar meus créditos",
        "tags": [
          "🔵 Créditos"
        ]
      }
    },
    "/v1/credits/balance/{userId}/initial": {
      "post": {
        "description": "Cria um saldo inicial de créditos para um usuário. Requer permissões de administrador.",
        "operationId": "CreditsController_createInitialCreditBalance",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "ID único do usuário",
            "schema": {
              "example": "123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Saldo inicial criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "userId": {
                      "type": "string",
                      "example": "123"
                    },
                    "balance": {
                      "type": "number",
                      "example": 0
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar saldo inicial de créditos",
        "tags": [
          "🔵 Créditos"
        ]
      }
    },
...
  },
  "info": {
    "title": "🧠 NeuralContent API",
    "description": "\n## API completa para plataforma de geração de conteúdo com IA\n\n### 🔐 Autenticação\nEsta API utiliza **JWT Bearer tokens**. Para autenticar suas requisições:\n```\nAuthorization: Bearer {seu_jwt_token}\n```\n\n### 📊 Rate Limiting\n- **Usuários Free:** 100 req/min\n- **Usuários Pro:** 500 req/min  \n- **Usuários Enterprise:** 1000 req/min\n- **Administradores:** Sem limite\n\n### 🔄 Versionamento\n- **Versão Atual:** v1.0.0\n- **URL Base:** `/v1`\n- **Suporte:** Versionamento semântico (SemVer)\n\n### 📈 Funcionalidades Principais\n- **Autenticação e Autorização** (JWT + Roles)\n- **Gestão de Usuários** (CRUD + Perfis)\n- **Sistema de Planos** (Free, Pro, Enterprise)\n- **Gestão de Créditos** (Compra, Consumo, Histórico)\n- **Processamento de Pagamentos** (Stripe Integration)\n- **Monitoramento e Logs** (Health Checks + Analytics)\n\n### 🛡️ Segurança\n- Criptografia de senhas (bcrypt)\n- Validação de entrada (class-validator)\n- Proteção CORS configurável\n- Headers de segurança (Helmet)\n- Rate limiting por usuário\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "NeuralContent Support",
      "url": "https://neuralcontent.com/support",
      "email": "api-support@neuralcontent.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://neuralcontent.com/terms"
  },
...
  "servers": [
    {
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "JWT token obtido via /auth/login. Formato: Bearer {token}",
        "in": "header"
      },
      "ApiKey-auth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key para integração de sistemas externos"
      }
    },
    "schemas": {
...
      "ValidateCreditsDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "ID do usuário",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "amount": {
            "type": "number",
            "description": "Quantidade de créditos necessários",
            "example": 50,
            "minimum": 1,
            "maximum": 10000
          },
          "serviceType": {
            "type": "string",
            "description": "Tipo do serviço para calcular custo",
            "example": "text_generation"
          }
        },
        "required": [
          "userId",
          "amount",
          "serviceType"
        ]
      },
      "ValidateCreditsResponseDto": {
        "type": "object",
        "properties": {
          "hasEnoughCredits": {
            "type": "boolean",
            "description": "Se o usuário tem créditos suficientes",
            "example": true
          },
          "currentBalance": {
            "type": "number",
            "description": "Saldo atual de créditos",
            "example": 1000
          },
          "requiredAmount": {
            "type": "number",
            "description": "Quantidade necessária",
            "example": 50
          },
          "serviceCost": {
            "type": "number",
            "description": "Custo do serviço calculado",
            "example": 100
          },
          "remaining": {
            "type": "number",
            "description": "Créditos restantes após a operação",
            "example": 900
          },
          "canProceed": {
            "type": "boolean",
            "description": "Se pode prosseguir com a operação",
            "example": true
          },
          "serviceType": {
            "type": "string",
            "description": "Tipo do serviço",
            "example": "text_generation"
          },
          "message": {
            "type": "string",
            "description": "Mensagem explicativa",
            "example": "Saldo suficiente para a operação"
          }
        },
        "required": [
          "hasEnoughCredits",
          "currentBalance",
          "requiredAmount",
          "serviceCost",
          "remaining",
          "canProceed",
          "serviceType",
          "message"
        ]
      },
      "ConsumeCreditsDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "ID do usuário que está consumindo os créditos",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "amount": {
            "type": "number",
            "description": "Quantidade de créditos a serem consumidos",
            "example": 50,
            "minimum": 1,
            "maximum": 10000
          },
          "serviceType": {
            "type": "string",
            "description": "Tipo de serviço que está consumindo os créditos",
            "example": "text_generation",
            "enum": [
              "text_generation",
              "image_generation",
              "translation",
              "summarization",
              "voice_synthesis",
              "document_analysis",
              "custom_service"
            ]
          },
          "description": {
            "type": "string",
            "description": "Descrição detalhada do consumo",
            "example": "Geração de texto com 1500 tokens usando modelo GPT-4",
            "maxLength": 500
          },
          "sessionId": {
            "type": "string",
            "description": "ID da sessão ou processo que originou o consumo",
            "example": "session_12345",
            "maxLength": 255
          },
          "clientIp": {
            "type": "string",
            "description": "IP do cliente",
            "example": "192.168.1.1"
          },
          "userAgent": {
            "type": "string",
            "description": "User agent do cliente",
            "example": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36"
          },
          "metadata": {
            "type": "object",
            "description": "Metadados adicionais do consumo",
            "example": {
              "requestId": "req_12345",
              "model": "gpt-4",
              "tokens": 1500,
              "duration": 2.5,
              "parameters": {
                "temperature": 0.7,
                "maxTokens": 2000
              }
            }
          }
        },
        "required": [
          "userId",
          "amount",
          "serviceType",
          "description"
        ]
      },
      "ConsumeCreditsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Se o consumo foi bem-sucedido",
            "example": true
          },
          "transactionId": {
            "type": "string",
            "description": "ID da transação criada",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "amountConsumed": {
            "type": "number",
            "description": "Quantidade de créditos consumidos",
            "example": 50
          },
          "newBalance": {
            "type": "number",
            "description": "Saldo de créditos após o consumo",
            "example": 950
          },
          "previousBalance": {
            "type": "number",
            "description": "Saldo anterior",
            "example": 1000
          },
          "serviceType": {
            "type": "string",
            "description": "Tipo do serviço consumido",
            "example": "text_generation"
          },
          "message": {
            "type": "string",
            "description": "Mensagem descritiva do resultado",
            "example": "Créditos consumidos com sucesso"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp do consumo",
            "example": "2025-07-14T10:15:00.000Z"
          }
        },
        "required": [
          "success",
          "transactionId",
          "amountConsumed",
          "newBalance",
          "previousBalance",
          "serviceType",
          "message",
          "timestamp"
        ]
      },
...
    }
  }
}